+++ packages/elements/user-picker/src/components/ClearIndicator.tsx
@@ -1,6 +1,7 @@
+import * as React from 'react';
 import { components } from '@atlaskit/select';
 import Tooltip from '@atlaskit/tooltip';
-import * as React from 'react';
+import { isPopupUserPickerByProps } from './utils';
 
 export class ClearIndicator extends React.PureComponent<any> {
   private handleMouseDown = (event: React.MouseEvent) => {
@@ -9,7 +10,11 @@ export class ClearIndicator extends React.PureComponent<any> {
     }
     this.props.clearValue();
     // Prevent focus when clear on blurred state
-    if (this.props.selectProps && !this.props.selectProps.isFocused) {
+    const { selectProps } = this.props;
+    if (
+      selectProps &&
+      (!selectProps.isFocused || isPopupUserPickerByProps(selectProps))
+    ) {
       event.stopPropagation();
     }
   };