+++ packages/elements/user-picker/src/index.ts
@@ -4,5 +4,6 @@ export {
   isValidEmail,
 } from './components/emailValidation';
 export { UserPicker as default } from './components/UserPicker';
+export { PopupUserPicker } from './components/PopupUserPicker';
 export { isEmail, isTeam, isUser } from './components/utils';
 export * from './types';